+++ code/src/app/components/page.tsx
@@ -1,15 +1,11 @@
-import Header from "./core/header"
 import { PageContent } from "../models/pageContent";
-export default function Page({ id, mainHeader, content }: PageContent) {
+export default function Page({ id, content }: PageContent) {
     return (
-
         <section
             id={id}
             className="h-screen snap-start snap-always flex flex-col"
         >
-            {mainHeader && <Header />}
             <div className="flex-grow flex flex-col m-4">
-              
                 <div>{content}</div>
             </div>
         </section>
