+++ code/src/app/domains/domains.tsx
@@ -1,0 +1,7 @@
+export default function Domains() {
+  return (
+    <div>
+      <h1>Domains</h1>
+    </div>
+  );
+}
