+++ code/src/app/landing/landing.tsx
@@ -1,0 +1,7 @@
+export default function Landing() {
+    return (
+        <div>
+            <h1>Landing</h1>
+        </div>
+    );
+}
