+++ code/src/app/models/pageContent.ts
@@ -3,7 +3,6 @@ import { type ReactElement } from "react";
 
 export interface PageContent {
   id: string;
-  mainHeader?: boolean;
   content: ReactElement;
 }
