+++ code/src/app/offerings/offerings.tsx
@@ -1,0 +1,7 @@
+export default function Offerings() {
+    return (
+        <div>
+            <h1>Offerings</h1>
+        </div>
+    );
+}
